monitor: Display human-readable identities
This commit changes cilium monitor's output to display human-readable
identities instead of numbers for reserved (fixed) identities.

The output before this commit ressembles:

  <- host flow 0x0 identity 2->0 state new ifindex enp0s8 orig-ip 0.0.0.0: 192.168.33.12:31313 -> 10.11.0.113:40364 tcp SYN, ACK

After:

  <- host flow 0x0 identity world->unknown state new ifindex enp0s8 orig-ip 0.0.0.0: 192.168.33.12:31313 -> 10.11.0.113:40364 tcp SYN, ACK

+++ pkg/hubble/parser/threefour/parser.go
@@ -463,16 +463,16 @@ func decodeSecurityIdentities(dn *monitor.DropNotify, tn *monitor.TraceNotify, p
 ) {
 	switch {
 	case dn != nil:
-		sourceSecurityIdentiy = dn.SrcLabel
-		destinationSecurityIdentity = dn.DstLabel
+		sourceSecurityIdentiy = uint32(dn.SrcLabel)
+		destinationSecurityIdentity = uint32(dn.DstLabel)
 	case tn != nil:
-		sourceSecurityIdentiy = tn.SrcLabel
-		destinationSecurityIdentity = tn.DstLabel
+		sourceSecurityIdentiy = uint32(tn.SrcLabel)
+		destinationSecurityIdentity = uint32(tn.DstLabel)
 	case pvn != nil:
 		if pvn.IsTrafficIngress() {
-			sourceSecurityIdentiy = pvn.RemoteLabel
+			sourceSecurityIdentiy = uint32(pvn.RemoteLabel)
 		} else {
-			destinationSecurityIdentity = pvn.RemoteLabel
+			destinationSecurityIdentity = uint32(pvn.RemoteLabel)
 		}
 	}
